fixed PR comments
