+++ src/client/extension.ts
@@ -56,8 +56,7 @@ export async function activate(
             } else if (
                 vscode.workspace.workspaceFolders !== undefined &&
                 PortalWebView.currentPanel &&
-                PortalWebView.currentDocument !==
-                vscode.window.activeTextEditor.document.fileName &&
+                PortalWebView.currentDocument !== vscode.window.activeTextEditor.document.fileName &&
                 PortalWebView.checkDocumentIsHTML()
             ) {
                 PortalWebView.currentPanel._update();
