+++ src/server/htmlServer.ts
@@ -17,9 +17,8 @@ import {
 } from 'vscode-languageserver/node';
 import * as nearley from 'nearley';
 import { URL } from 'url';
-
-import { getEditedLineContent, getMatchedManifestRecords, IManifestElement } from './lib/PortalAutoCompleteHelper';
-
+import { getEditedLineContent } from './lib/LineReader';
+import { getMatchedManifestRecords, IManifestElement } from './lib/PortalManifestReader';
 import {
     TextDocument
 } from 'vscode-languageserver-textdocument';
